DOCS-#2996: move doc-decorators to a separate file
