+++ modin/backends/base/__init__.py
@@ -11,6 +11,8 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
+"""The module represents the base query compiler that defines the common query compiler API."""
+
 from .query_compiler import BaseQueryCompiler
 
 __all__ = ["BaseQueryCompiler"]
